Some docs, and factor out modular change encoding
