+++ packages/dds/tree/src/feature-libraries/modular-schema/genericFieldKind.ts
@@ -18,10 +18,16 @@ import {
 import { FieldKind, Multiplicity } from "./fieldKind";
 
 /**
- * A field-agnostic change to a single element of a field.
+ * A field-kind-agnostic change to a single node within a field.
  */
 export interface GenericChange {
+    /**
+     * Index withing the field of the changed node.
+     */
     index: number;
+    /**
+     * Change to the node.
+     */
     nodeChange: NodeChangeset;
 }
 
@@ -30,6 +36,7 @@ export interface GenericChange {
  */
 export interface EncodedGenericChange {
     index: number;
+    // TODO: this format needs more documentation (ideally in the form of more specific types).
     nodeChange: JsonCompatibleReadOnly;
 }
 
@@ -120,11 +127,9 @@ export const genericChangeHandler: FieldChangeHandler<GenericChangeset> = {
             change: GenericChangeset,
             encodeChild: NodeChangeEncoder,
         ): JsonCompatibleReadOnly {
-            // Would use `change.map(...)` but the type system doesn't accept it
-            const encoded: JsonCompatibleReadOnly[] & EncodedGenericChangeset = [];
-            for (const { index, nodeChange } of change) {
-                encoded.push({ index, nodeChange: encodeChild(nodeChange) });
-            }
+            const encoded: JsonCompatibleReadOnly[] & EncodedGenericChangeset = change.map(
+                ({ index, nodeChange }) => ({ index, nodeChange: encodeChild(nodeChange) }),
+            );
             return encoded;
         },
         decodeJson: (
